remove plot_ens files
