+++ training/tests/integration/config/test_temporal_downscaler_ensemble.yaml
@@ -29,7 +29,7 @@ diagnostics:
     parameters: [tp]
     callbacks:
       # Add plot callbacks here.
-      - _target_: anemoi.training.diagnostics.callbacks.plot_ens.PlotEnsSample
+      - _target_: anemoi.training.diagnostics.callbacks.plot.PlotEnsSample
         dataset_names: ["data"]
         sample_idx: ${diagnostics.plot.sample_idx}
         per_sample : 2
@@ -39,24 +39,24 @@ diagnostics:
         accumulation_levels_plot: [0, 0.05, 0.1, 0.25, 0.5, 1, 1.5, 2, 3, 4, 5, 6, 7, 100] # in mm
         members: null  # None for all members, list for specific members
 
-      - _target_: anemoi.training.diagnostics.callbacks.plot_ens.PlotLoss
+      - _target_: anemoi.training.diagnostics.callbacks.plot.PlotLoss
         dataset_names: ["data"]
         parameter_groups:
           moisture: [tp, cp, tcw]
           sfc_wind: [10u, 10v]
         every_n_batches: ${diagnostics.plot.frequency.batch}
         #! DEACTIVATING UNTIL PYSHTOOLS FIX FOR 3.13
-      # - _target_: anemoi.training.diagnostics.callbacks.plot_ens.PlotSpectrum
+      # - _target_: anemoi.training.diagnostics.callbacks.plot.PlotSpectrum
       #   dataset_names: ["data"]
       #   sample_idx: ${diagnostics.plot.sample_idx}
       #   parameters: ${diagnostics.plot.parameters}
       #   every_n_batches: ${diagnostics.plot.frequency.batch}
-      - _target_: anemoi.training.diagnostics.callbacks.plot_ens.PlotHistogram
+      - _target_: anemoi.training.diagnostics.callbacks.plot.PlotHistogram
         dataset_names: ["data"]
         sample_idx: ${diagnostics.plot.sample_idx}
         parameters: ${diagnostics.plot.parameters}
         every_n_batches: ${diagnostics.plot.frequency.batch}
         precip_and_related_fields: ["tp", "cp"] # Optional: specify precip fields for special histogram treatment
-      - _target_: anemoi.training.diagnostics.callbacks.plot_ens.GraphTrainableFeaturesPlot
+      - _target_: anemoi.training.diagnostics.callbacks.plot.GraphTrainableFeaturesPlot
         dataset_names: ["data"]
         every_n_epochs: ${diagnostics.plot.frequency.epoch}
